pyproject.toml
[project]
name = "izthere"
version = "0.1.0"
description = "Simple modular monitor/notifier for the web"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "playwright>=1.58.0",
    "python-telegram-bot>=22.6",
    "pyyaml>=6.0.3",
]

[tool.uv.build-backend]
packages = ["is"]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project.scripts]
izthere = "izthere.main:main"

[tool.pyright]
reportExplicitAny = false
reportAny = false
reportUnknownArgumentType = false
